+++ returns/maybe.py
@@ -1,4 +1,4 @@
-from abc import ABCMeta
+from abc import ABCMeta, abstractmethod
 from functools import wraps
 from typing import (
     Any,
@@ -16,7 +16,7 @@ from typing import (
 from typing_extensions import final
 
 from returns._generated.iterable import iterable
-from returns.interfaces import applicative, bindable, mappable
+from returns.interfaces import applicative, bindable, mappable, unwrappable
 from returns.primitives.container import BaseContainer
 from returns.primitives.exceptions import UnwrapFailedError
 from returns.primitives.hkt import Kind1
@@ -36,6 +36,7 @@ class Maybe(
     mappable.Mappable1[_ValueType],
     bindable.Bindable1[_ValueType],
     applicative.Applicative1[_ValueType],
+    unwrappable.Unwrappable[_ValueType, None],
     metaclass=ABCMeta,
 ):
     """
@@ -59,6 +60,7 @@ class Maybe(
     #: Failure type that is used to represent the failed computation.
     failure_type: ClassVar[Type['_Nothing']]
 
+    @abstractmethod  # noqa: WPS125
     def map(  # noqa: WPS125
         self,
         function: Callable[[_ValueType], Optional[_NewValueType]],
@@ -76,8 +78,8 @@ class Maybe(
           >>> assert Nothing.map(mappable) == Nothing
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def apply(
         self,
         function: Kind1['Maybe', Callable[[_ValueType], _NewValueType]],
@@ -98,8 +100,8 @@ class Maybe(
           >>> assert Nothing.apply(Nothing) == Nothing
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def bind(
         self,
         function: Callable[[_ValueType], Kind1['Maybe', _NewValueType]],
@@ -117,7 +119,6 @@ class Maybe(
           >>> assert Nothing.bind(bindable) == Nothing
 
         """
-        raise NotImplementedError
 
     def value_or(
         self,
@@ -133,8 +134,8 @@ class Maybe(
           >>> assert Nothing.value_or(1) == 1
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def or_else_call(
         self,
         function: Callable[[], _NewValueType],
@@ -169,8 +170,8 @@ class Maybe(
           ValueError: Nothing!
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def unwrap(self) -> _ValueType:
         """
         Get value from successful container or raise exception for failed one.
@@ -186,8 +187,8 @@ class Maybe(
           returns.primitives.exceptions.UnwrapFailedError
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def failure(self) -> None:
         """
         Get failed value from failed container or raise exception from success.
@@ -203,7 +204,6 @@ class Maybe(
           returns.primitives.exceptions.UnwrapFailedError
 
         """
-        raise NotImplementedError
 
     @classmethod
     def from_value(