+++ returns/result.py
@@ -1,4 +1,4 @@
-from abc import ABCMeta
+from abc import ABCMeta, abstractmethod
 from functools import wraps
 from inspect import FrameInfo
 from typing import (
@@ -18,7 +18,7 @@ from typing import (
 from typing_extensions import final
 
 from returns._generated.iterable import iterable
-from returns.interfaces import applicative, bindable, mappable
+from returns.interfaces import applicative, bindable, mappable, unwrappable
 from returns.primitives.container import BaseContainer
 from returns.primitives.exceptions import UnwrapFailedError
 from returns.primitives.hkt import Kind2
@@ -40,6 +40,7 @@ class Result(
     mappable.Mappable2[_ValueType, _ErrorType],
     bindable.Bindable2[_ValueType, _ErrorType],
     applicative.Applicative2[_ValueType, _ErrorType],
+    unwrappable.Unwrappable[_ValueType, _ErrorType],
     metaclass=ABCMeta,
 ):
     """
@@ -70,6 +71,7 @@ class Result(
         """Returns a list with stack trace when :func:`~Failure` was called."""
         return self._trace
 
+    @abstractmethod  # noqa: WPS125
     def map(  # noqa: WPS125
         self,
         function: Callable[[_ValueType], _NewValueType],
@@ -88,8 +90,8 @@ class Result(
           >>> assert Failure('a').map(mappable) == Failure('a')
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def apply(
         self,
         container: Kind2[
@@ -115,8 +117,8 @@ class Result(
           >>> assert isinstance(with_failure, Result.failure_type)
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def bind(
         self,
         function: Callable[
@@ -141,7 +143,6 @@ class Result(
           >>> assert Failure('a').bind(bindable) == Failure('a')
 
         """
-        raise NotImplementedError
 
     def unify(
         self,
@@ -176,6 +177,7 @@ class Result(
         """
         return self.bind(function)  # type: ignore
 
+    @abstractmethod
     def fix(
         self,
         function: Callable[[_ErrorType], _NewValueType],
@@ -194,8 +196,8 @@ class Result(
           >>> assert Failure('a').fix(fixable) == Success('ab')
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def alt(
         self,
         function: Callable[[_ErrorType], _NewErrorType],
@@ -214,8 +216,8 @@ class Result(
           >>> assert Failure('a').alt(altable) == Failure('ab')
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def rescue(
         self,
         function: Callable[
@@ -239,8 +241,8 @@ class Result(
           >>> assert Failure('aa').rescue(rescuable) == Success('aab')
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def value_or(
         self,
         default_value: _NewValueType,
@@ -255,8 +257,8 @@ class Result(
           >>> assert Failure(1).value_or(2) == 2
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def unwrap(self) -> _ValueType:
         """
         Get value or raise exception.
@@ -272,8 +274,8 @@ class Result(
           returns.primitives.exceptions.UnwrapFailedError
 
         """
-        raise NotImplementedError
 
+    @abstractmethod
     def failure(self) -> _ErrorType:
         """
         Get failed value or raise exception.
@@ -289,7 +291,6 @@ class Result(
           returns.primitives.exceptions.UnwrapFailedError
 
         """
-        raise NotImplementedError
 
     @classmethod
     def from_value(