+++ tests/test_io/test_ioresult_container/test_ioresult_base.py
@@ -1,18 +1,4 @@
-import pytest
-
 from returns.io import IOResult
-
-
-@pytest.mark.parametrize('method_name', [
-    'bind',
-    'bind_result',
-    'rescue',
-])
-def test_ioresult_abstract_method(method_name):
-    """Checks that Result itself contains abstract methods."""
-    method = getattr(IOResult, method_name)
-    with pytest.raises(NotImplementedError):
-        method(object, lambda to_output: to_output)
 
 
 def test_ioresult_types():